app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
ocrmypdf==16.1.0
PyMuPDF==1.23.17
pdf2image==3.0.1
Pillow==10.1.0
requests==2.31.0
numpy==1.25.2
sentence-transformers==2.2.2
ollama==0.1.7
pydantic==2.5.0
python-dotenv==1.0.0